Move parse_md_file tests into test_config.py; delete test_frontmatter.py
